Create README.md
Hello! 👋
I'm Alvaro Anaya, a passionate student of Data Analysis with a focus on the fascinating world of data science. I'm from Spain, and currently, I'm immersed in a Data Analysis course that has provided me with a solid foundation in statistics, programming, and data visualization.
About Me
🌱 I am currently learning and honing my skills in Python, SQL, and popular data science tools such as Pandas, NumPy, and Matplotlib.
👀 I am interested in exploring datasets, uncovering patterns, and telling compelling stories through data.
💞️ I am open to collaborations and interesting projects related to data analysis.
Contact
📫 You can find me on GitHub as @AlvaroAnaya24DAnalysis.
I am excited about the potential of the data science world and eager to contribute to innovative projects! If you have any ideas or want to collaborate, feel free to get in touch with me!